requirements.txt
Flask>=3.0
python-dotenv>=1.0
Werkzeug>=3.0
Pillow>=10.0
